Nullpointer check in case skeleton generator is not provided.
git-svn-id: https://svn.apache.org/repos/asf/incubator/openejb/trunk@445044 13f79535-47bb-0310-9956-ffa450edef68

+++ openejb2/modules/openejb-builder/src/java/org/openejb/deployment/OpenEJBModuleBuilder.java
@@ -301,19 +301,21 @@ public class OpenEJBModuleBuilder implements ModuleBuilder, EJBReferenceBuilder 
         entityBuilder.initContext(earContext, moduleJ2eeContext, moduleUri, cl, enterpriseBeans, interfaces);
         mdbBuilder.initContext(cl, enterpriseBeans);
 
-        File tempJar = null;
-        try {
-            tempJar = DeploymentUtil.createTempFile();
-
-            skeletonGenerator.generateSkeletons(interfaces, tempJar, cl);
-
-            earContext.addIncludeAsPackedJar(URI.create("corba.jar"), new JarFile(tempJar));
-        } catch (IOException e) {
-            throw new DeploymentException("Unable to generate CORBA skels for: " + moduleUri, e);
-        } catch (CompilerException e) {
-            throw new DeploymentException("Unable to generate CORBA skels for: " + moduleUri, e);
-        } finally {
-            tempJar.delete();
+        if( skeletonGenerator!=null ) {
+	        File tempJar = null;
+	        try {
+	            tempJar = DeploymentUtil.createTempFile();
+	
+	            skeletonGenerator.generateSkeletons(interfaces, tempJar, cl);
+	
+	            earContext.addIncludeAsPackedJar(URI.create("corba.jar"), new JarFile(tempJar));
+	        } catch (IOException e) {
+	            throw new DeploymentException("Unable to generate CORBA skels for: " + moduleUri, e);
+	        } catch (CompilerException e) {
+	            throw new DeploymentException("Unable to generate CORBA skels for: " + moduleUri, e);
+	        } finally {
+	            tempJar.delete();
+	        }
         }
     }
